Controller, Services and Tests done
Frontend able to be implemented propertly

+++ backend/quizzgame/src/main/java/com/chalet/quizzgame/repository/QuestionRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface QuestionRepository extends JpaRepository<Question, Long>{
+public interface QuestionRepository extends JpaRepository<Question, Integer>{
 
     Optional<Question> findByQuestion(String question);
 
-    Optional<Question> findByQuestionId(Long Id);
+    Optional<Question> findByQuestionId(Integer Id);
 
 }
